LayoutTests/media/video-playback-restriction-play-before-load-expected.txt

Test that, when RequireUserGestureForAudioRateChange is set, starting playback of an audio file requires a user gesture.


RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForAudioRateChange"))
RUN(video.src = findMediaFile('video', 'content/test'))
RUN(video.play().then(resolvePromise, rejectPromise))
Promise correctly rejected.
END OF TEST
